+++ _test_common/lib/test_phases.dart
@@ -11,8 +11,7 @@ import 'package:build_test/build_test.dart';
 import 'package:logging/logging.dart';
 import 'package:test/test.dart';
 
-import 'in_memory_reader.dart';
-import 'in_memory_writer.dart';
+import 'in_memory_reader_writer.dart';
 import 'package_graphs.dart';
 
 Future<void> wait(int milliseconds) =>
@@ -41,13 +40,12 @@ void _printOnFailure(LogRecord record) {
 /// For example `$$myapp|lib/utils.copy.dart` will check that the generated
 /// output was written to the build cache.
 ///
+/// [resumeFrom] reuses the `readerWriter` from a previous [BuildResult].
+///
 /// [packageGraph] supplies the root package into which the outputs are to be
 /// written.
 ///
 /// [status] optionally indicates the desired outcome.
-///
-/// [writer] can optionally be provided to capture assets written by the
-/// builders (e.g. when [outputs] is not sufficient).
 ///
 /// [logLevel] sets the builder log level and [onLog] can optionally capture
 /// build log messages.
@@ -67,15 +65,14 @@ void _printOnFailure(LogRecord record) {
 ///       });
 ///     }
 ///
-Future<BuildResult> testBuilders(
+Future<TestBuildersResult> testBuilders(
   List<BuilderApplication> builders,
   Map<String, /*String|List<int>*/ Object> inputs, {
+  TestBuildersResult? resumeFrom,
   Map<String, /*String|List<int>*/ Object>? outputs,
   PackageGraph? packageGraph,
   BuildStatus status = BuildStatus.success,
   Map<String, BuildConfig>? overrideBuildConfig,
-  InMemoryRunnerAssetReader? reader,
-  InMemoryRunnerAssetWriter? writer,
   Level? logLevel,
   // A better way to "silence" logging than setting logLevel to OFF.
   void Function(LogRecord record) onLog = _printOnFailure,
@@ -88,14 +85,16 @@ Future<BuildResult> testBuilders(
   Set<BuildFilter> buildFilters = const {},
   String? logPerformanceDir,
   String expectedGeneratedDir = 'generated',
+  void Function(AssetId id)? onDelete,
 }) async {
   packageGraph ??= buildPackageGraph({rootPackage('a'): []});
-  writer ??= InMemoryRunnerAssetWriter();
-  reader ??= InMemoryRunnerAssetReader.shareAssetCache(writer.assets,
-      rootPackage: packageGraph.root.name);
+  final readerWriter = resumeFrom == null
+      ? InMemoryRunnerAssetReaderWriter(rootPackage: packageGraph.root.name)
+      : resumeFrom.readerWriter;
+  readerWriter.onDelete = onDelete;
   var pkgConfigId =
       AssetId(packageGraph.root.name, '.dart_tool/package_config.json');
-  if (!await reader.canRead(pkgConfigId)) {
+  if (!await readerWriter.canRead(pkgConfigId)) {
     var packageConfig = {
       'configVersion': 2,
       'packages': [
@@ -108,21 +107,21 @@ Future<BuildResult> testBuilders(
           },
       ],
     };
-    await writer.writeAsString(pkgConfigId, jsonEncode(packageConfig));
+    await readerWriter.writeAsString(pkgConfigId, jsonEncode(packageConfig));
   }
 
   inputs.forEach((serializedId, contents) {
     var id = makeAssetId(serializedId);
     if (contents is String) {
-      reader!.cacheStringAsset(id, contents);
+      readerWriter.cacheStringAsset(id, contents);
     } else if (contents is List<int>) {
-      reader!.cacheBytesAsset(id, contents);
+      readerWriter.cacheBytesAsset(id, contents);
     }
   });
 
   builderConfigOverrides ??= const {};
   var environment = OverrideableEnvironment(IOEnvironment(packageGraph),
-      reader: reader, writer: writer, onLog: onLog);
+      reader: readerWriter, writer: readerWriter, onLog: onLog);
   var logSubscription =
       LogSubscription(environment, verbose: verbose, logLevel: logLevel);
   var options = await BuildOptions.create(logSubscription,
@@ -149,12 +148,13 @@ Future<BuildResult> testBuilders(
   if (checkBuildStatus) {
     checkBuild(result,
         outputs: outputs,
-        writer: writer,
+        writer: readerWriter,
         status: status,
         rootPackage: packageGraph.root.name,
         expectedGeneratedDir: expectedGeneratedDir);
   }
-  return result;
+
+  return TestBuildersResult(buildResult: result, readerWriter: readerWriter);
 }
 
 /// Translates expected outptus which start with `$$` to the build cache and
@@ -189,4 +189,11 @@ void checkBuild(BuildResult result,
     checkOutputs(unhiddenOutputs, result.outputs, writer,
         mapAssetIds: (id) => mapHidden(id, expectedGeneratedDir));
   }
+}
+
+class TestBuildersResult {
+  final BuildResult buildResult;
+  final InMemoryRunnerAssetReaderWriter readerWriter;
+
+  TestBuildersResult({required this.buildResult, required this.readerWriter});
 }