+++ build_modules/test/util.dart
@@ -17,15 +17,13 @@ Future<void> testBuilderAndCollectAssets(
     void Function(LogRecord log)? onLog,
     void Function(AssetId, Iterable<AssetId>)?
         reportUnusedAssetsForInput}) async {
-  var writer = InMemoryAssetWriter();
   onLog ??= (log) => printOnFailure('${log.level}: ${log.message}');
-  await testBuilder(builder, assets,
+  final result = await testBuilder(builder, assets,
       generateFor: generateFor,
       outputs: outputs,
       onLog: onLog,
-      reportUnusedAssetsForInput: reportUnusedAssetsForInput,
-      writer: writer);
-  writer.assets.forEach((id, value) {
+      reportUnusedAssetsForInput: reportUnusedAssetsForInput);
+  result.readerWriter.assets.forEach((id, value) {
     assets['${id.package}|${id.path}'] = value;
   });
 }