+++ build_test/test/test_builder_test.dart
@@ -75,23 +75,6 @@ void main() {
     }
   });
 
-  test('can pass a custom reader', () async {
-    var reader =
-        await PackageAssetReader.currentIsolate(rootPackage: 'build_test');
-    var builder = TestBuilder(
-        buildExtensions: {
-          '.txt': ['.txt.copy']
-        },
-        build: (buildStep, _) async {
-          await buildStep.writeAsString(
-              buildStep.inputId.addExtension('.copy'),
-              buildStep
-                  .readAsString(AssetId('build_test', 'test/data/hello.txt')));
-        });
-    await testBuilder(builder, {'build_test|data/1.txt': ''},
-        outputs: {'build_test|data/1.txt.copy': 'hello world'}, reader: reader);
-  });
-
   test('can capture reportUnusedAssets calls', () async {
     var unusedInput = AssetId('a', 'lib/unused.txt');
     var recorded = <AssetId, Iterable<AssetId>>{};