+++ build_web_compilers/test/util.dart
@@ -11,11 +11,9 @@ import 'package:test/test.dart';
 /// Forwards to [testBuilder], and adds all output assets to [assets].
 Future<void> testBuilderAndCollectAssets(
     Builder builder, Map<String, Object> assets) async {
-  var writer = InMemoryAssetWriter();
-  await testBuilder(builder, assets,
-      writer: writer,
+  final result = await testBuilder(builder, assets,
       onLog: (log) => printOnFailure('${log.level}: ${log.message}'));
-  writer.assets.forEach((id, value) {
+  result.readerWriter.assets.forEach((id, value) {
     assets['${id.package}|${id.path}'] = value;
   });
 }